requirement.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.121.3
fonttools==4.60.1
h11==0.16.0
idna==3.11
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
numpy==2.3.4
packaging==25.0
pillow==12.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
rich==14.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
